+++ src/main/scala/uk/gov/nationalarchives/tdr/api/auth/ValidationAuthoriser.scala
@@ -3,7 +3,6 @@ package uk.gov.nationalarchives.tdr.api.auth
 import sangria.execution.{BeforeFieldResult, Middleware, MiddlewareAttachment, MiddlewareBeforeField, MiddlewareQueryContext}
 import sangria.schema.Context
 import uk.gov.nationalarchives.tdr.api.graphql.ConsignmentApiContext
-import uk.gov.nationalarchives.tdr.api.graphql.Tags._
 object ValidationAuthoriser extends Middleware[ConsignmentApiContext] with MiddlewareBeforeField[ConsignmentApiContext] {
 
   case class AuthorisationException(message: String) extends Exception(message)
@@ -18,7 +17,7 @@ object ValidationAuthoriser extends Middleware[ConsignmentApiContext] with Middl
   override def beforeField(queryVal: QueryVal, mctx: MiddlewareQueryContext[ConsignmentApiContext, _, _]
                            , ctx: Context[ConsignmentApiContext, _]): BeforeFieldResult[ConsignmentApiContext, Unit] = {
       val validationList: Seq[BeforeFieldResult[ConsignmentApiContext, Unit]] = ctx.field.tags.map {
-        case v: ValidateTags => v.validate(ctx)
+        case v: AuthorisationTag => v.validate(ctx)
       }
     validationList.headOption.getOrElse(continue)
   }
